@jakirkham 2021-10-21
Also PR ( https://github.com/MagicStack/uvloop/pull/445 ) may be relevant for us when sending multiple buffers