@jcrist 2021-11-05
Ok, this all works now and is _generally_ (with some caveats) faster than the existing tornado comms. It's surprisingly hard to isolate the comms from everything else (e.g. the protocol) - here's the benchmark I've been using:

<details>
<summary>Benchmark Script</summary>

```python
import argparse
import os
from dask.utils import format_time, format_bytes
from distributed.comm import listen, connect, CommClosedError
from distributed.protocol.serialize import Serialized
import asyncio, time


async def handle_comm(comm):
    try:
        while True:
            msg = await comm.read()
            await comm.write(msg=msg)
    except CommClosedError:
        pass


class Bench:
    def __init__(self, n_bytes, n_frames, seconds):
        self.n_bytes = n_bytes
        self.n_frames = n_frames
        self.seconds = seconds
        self.running = True

    async def _loop(self):
        frames = [os.urandom(self.n_bytes) for _ in range(self.n_frames)]
        msg = Serialized({}, frames)
        async with listen("tcp://127.0.0.1:8080", handle_comm, deserialize=False):
            comm = await connect("tcp://127.0.0.1:8080", deserialize=False)
            start = time.time()
            nmsgs = 0
            while self.running:
                await comm.write(msg=msg)
                await comm.read()
                nmsgs += 1
            end = time.time()
            print(f"{format_time((end - start) / nmsgs)} roundtrip")
            await comm.close()

    async def _stopper(self):
        await asyncio.sleep(self.seconds)
        self.running = False

    async def entry(self):
        asyncio.create_task(self._stopper())
        await self._loop()

    def run(self):
        asyncio.get_event_loop().run_until_complete(self.entry())


def main():
    parser = argparse.ArgumentParser(description="Benchmark channels")
    parser.add_argument(
        "--frames",
        "-f",
        default=1,
        type=int,
        help="Number of frames per message",
    )
    parser.add_argument(
        "--bytes", "-b", default=1000, type=float, help="total payload size in bytes"
    )
    parser.add_argument(
        "--seconds", "-s", default=5, type=int, help="bench duration in secs"
    )
    parser.add_argument("--uvloop", action="store_true", help="Whether to use uvloop")
    args = parser.parse_args()

    if args.uvloop:
        import uvloop

        uvloop.install()

    print(
        f"Benchmarking -- n_frames = {args.frames}, "
        f"bytes_per_frame = {format_bytes(args.bytes)}, "
        f"uvloop={args.uvloop}"
    )

    Bench(int(args.bytes), args.frames, args.seconds).run()


if __name__ == "__main__":
    main()
```

</details>

A summary of results:

### Tornado based comms

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -f 1 -b 1e6
Benchmarking -- n_frames = 1, bytes_per_frame = 0.95 MiB, uvloop=False
842.27 us roundtrip
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -f 1 -b 1e5
Benchmarking -- n_frames = 1, bytes_per_frame = 97.66 kiB, uvloop=False
318.96 us roundtrip
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -f 1 -b 1e4
Benchmarking -- n_frames = 1, bytes_per_frame = 9.77 kiB, uvloop=False
265.72 us roundtrip
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -f 1 -b 1e3
Benchmarking -- n_frames = 1, bytes_per_frame = 0.98 kiB, uvloop=False
262.93 us roundtrip
```

### Asyncio based comms

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -f 1 -b 1e6
Benchmarking -- n_frames = 1, bytes_per_frame = 0.95 MiB, uvloop=False
780.15 us roundtrip
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -f 1 -b 1e5
Benchmarking -- n_frames = 1, bytes_per_frame = 97.66 kiB, uvloop=False
192.18 us roundtrip
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -f 1 -b 1e4
Benchmarking -- n_frames = 1, bytes_per_frame = 9.77 kiB, uvloop=False
151.51 us roundtrip
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -f 1 -b 1e3
Benchmarking -- n_frames = 1, bytes_per_frame = 0.98 kiB, uvloop=False
114.16 us roundtrip
```

So we're generally faster for single frame messages.

However! When writing multiple frames per message, we're slower:

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -f 5 -b 1e6
Benchmarking -- n_frames = 5, bytes_per_frame = 0.95 MiB, uvloop=False
3.79 ms roundtrip
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -f 5 -b 1e6
Benchmarking -- n_frames = 5, bytes_per_frame = 0.95 MiB, uvloop=False
4.49 ms roundtrip
```

I've tracked this down to this line here: https://github.com/dask/distributed/blob/06ba74b4e5e469d0b5d80f2b6dfc8b585585ed66/distributed/comm/tcp.py#L208.

The tornado based comms are allocating a single large array for all frames in a message, then filling that in. In the asyncio comm, I allocate a separate array for each frame. The allocation overhead of creating multiple `bytearray` objects is the cause of the slowdown here (`N` calls to malloc are slower than 1 large call), not anything to do with the IO pattern. If I move to a single allocation in the asyncio comms, we're faster again. But I don't think that's a good idea. Allocating one large array for all frames means that the lifetime of all frames are tied together - no frame can be freed until they all can. This can mean that memory will hang around a lot longer than expected. This may be a source of memory issues in distributed (please correct me if I'm wrong).

Note that to get the perf on write we had to muck with asyncio's internals. The default socket transport in asyncio will make unnecessary copies on write (hence the `_ZeroCopyWriter` wrapper class). This isn't needed with uvloop (where those copies don't happen), but is needed (but not implemented) on windows. I can fix this for windows if we want, just haven't put in the work yet. Also note that this fix as currently written won't be useful in a `TLS` context on asyncio without uvloop, since it wraps the transport rather than replacing it. Since openssl requires data copies anyway, I suspect this won't matter much, but haven't done enough benchmarking yet to be sure.