@gjoseph92 2021-11-05
Just curious, what are the results with `--uvloop`? I'm curious to see both overall performance, and which of tornado vs asyncio is better helped by uvloop out of the box.

> If I move to a single allocation in the asyncio comms, we're faster again. But I don't think that's a good idea. Allocating one large array for all frames means that the lifetime of all frames are tied together - no frame can be freed until they all can. This can mean that memory will hang around a lot longer than expected.

This is interesting, and at least deserves an xref to https://github.com/dask/distributed/pull/5208. @jakirkham can speak to this a lot more, but because we currently split frames in our protocol, often 1 frame < 1 output object that needs contiguous memory (such as a NumPy array). So we do want each block of _sub_-frames to all use the same contiguous memory, so we can zero-copy deserialize them.
However, when there are multiple logical frames/objects per message, then I agree that they should get separate buffers. This seems like yet another argument for not splitting frames at the protocol level, and instead letting the transport handle that, as you've mentioned in the past.