@jcrist 2021-11-08
Ok, here's a more complete benchmark script. It now supports multiple clients, enabling/disabling uvloop, and enabling/disabling TLS.

<details>
<summary>Benchmark Script</summary>

```python
import argparse
import asyncio
import datetime
import os
from concurrent.futures import ProcessPoolExecutor

from distributed.comm import listen, connect, CommClosedError
from distributed.protocol.serialize import Serialized


DIR = os.path.abspath(os.path.dirname(__file__))
CERT = os.path.join(DIR, "bench-cert.pem")
KEY = os.path.join(DIR, "bench-key.pem")


def ensure_certs():
    if not (os.path.exists(KEY) and os.path.exists(CERT)):
        from cryptography import x509
        from cryptography.hazmat.backends import default_backend
        from cryptography.hazmat.primitives import hashes
        from cryptography.hazmat.primitives import serialization
        from cryptography.hazmat.primitives.asymmetric import rsa
        from cryptography.x509.oid import NameOID

        key = rsa.generate_private_key(
            public_exponent=65537, key_size=2048, backend=default_backend()
        )
        key_bytes = key.private_bytes(
            encoding=serialization.Encoding.PEM,
            format=serialization.PrivateFormat.PKCS8,
            encryption_algorithm=serialization.NoEncryption(),
        )

        subject = issuer = x509.Name(
            [x509.NameAttribute(NameOID.COMMON_NAME, "ery-bench")]
        )
        now = datetime.datetime.utcnow()
        cert = (
            x509.CertificateBuilder()
            .subject_name(subject)
            .issuer_name(issuer)
            .public_key(key.public_key())
            .serial_number(x509.random_serial_number())
            .not_valid_before(now)
            .not_valid_after(now + datetime.timedelta(days=365))
            .sign(key, hashes.SHA256(), default_backend())
        )
        cert_bytes = cert.public_bytes(serialization.Encoding.PEM)
        with open(CERT, "wb") as f:
            f.write(cert_bytes)
        with open(KEY, "wb") as f:
            f.write(key_bytes)


def get_ssl_context():
    import ssl

    context = ssl.SSLContext(ssl.PROTOCOL_TLS)
    context.load_cert_chain(CERT, KEY)
    context.check_hostname = False
    context.verify_mode = ssl.CERT_NONE
    return context


class BenchProc:
    def __init__(self, n_bytes, n_frames, n_seconds, n_clients, use_tls):
        self.n_bytes = n_bytes
        self.n_frames = n_frames
        self.n_seconds = n_seconds
        self.n_clients = n_clients
        self.use_tls = use_tls

    async def client(self, address, comm_kwargs, msg):
        comm = await connect(address, deserialize=False, **comm_kwargs)

        while self.running:
            await comm.write(msg)
            await comm.read()

        await comm.close()

    def stop(self):
        self.running = False

    async def run(self):
        self.running = True
        if self.use_tls:
            kwargs = {"ssl_context": get_ssl_context()}
            prefix = "tls"
        else:
            kwargs = {}
            prefix = "tcp"

        address = f"{prefix}://127.0.0.1:8080"
        msg = Serialized({}, [os.urandom(self.n_bytes) for _ in range(self.n_frames)])

        loop = asyncio.get_running_loop()
        loop.call_later(self.n_seconds, self.stop)
        tasks = [
            asyncio.create_task(self.client(address, kwargs, msg))
            for _ in range(self.n_clients)
        ]
        await asyncio.gather(*tasks, return_exceptions=True)


def bench_proc_main(n_bytes, n_frames, n_seconds, n_clients, use_tls):
    bench = BenchProc(n_bytes, n_frames, n_seconds, n_clients, use_tls)
    asyncio.run(bench.run())


async def run(n_bytes, n_frames, n_seconds, n_procs, n_clients, use_tls):
    if use_tls:
        ensure_certs()
        kwargs = {"ssl_context": get_ssl_context()}
        prefix = "tls"
    else:
        kwargs = {}
        prefix = "tcp"

    address = f"{prefix}://127.0.0.1:8080"

    loop = asyncio.get_running_loop()

    count = 0

    async def handle_comm(comm):
        nonlocal count
        try:
            while True:
                msg = await comm.read()
                await comm.write(msg=msg)
                count += 1
        except CommClosedError:
            pass

    async with listen(address, handle_comm, deserialize=False, **kwargs):
        with ProcessPoolExecutor(max_workers=n_procs) as executor:
            tasks = [
                loop.run_in_executor(
                    executor,
                    bench_proc_main,
                    n_bytes,
                    n_frames,
                    n_seconds,
                    n_clients,
                    use_tls,
                )
                for _ in range(n_procs)
            ]
            await asyncio.gather(*tasks)

    print(f"{count / n_seconds} RPS")
    print(f"{n_seconds / count * 1e6} us per request")
    print(f"{n_bytes * count / (n_seconds * 1e6)} MB/s each way")


def main():
    parser = argparse.ArgumentParser(description="Benchmark channels")
    parser.add_argument(
        "--procs", "-p", default=1, type=int, help="Number of client processes"
    )
    parser.add_argument(
        "--concurrency",
        "-c",
        default=1,
        type=int,
        help="Number of clients per process",
    )
    parser.add_argument(
        "--frames",
        "-f",
        default=1,
        type=int,
        help="Number of frames per message",
    )
    parser.add_argument(
        "--bytes", "-b", default=1000, type=float, help="total payload size in bytes"
    )
    parser.add_argument(
        "--seconds", "-s", default=5, type=int, help="bench duration in secs"
    )
    parser.add_argument("--uvloop", action="store_true", help="Whether to use uvloop")
    parser.add_argument("--tls", action="store_true", help="Whether to use TLS")
    args = parser.parse_args()

    if args.uvloop:
        import uvloop

        uvloop.install()

    n_bytes = int(args.bytes)

    print(
        f"processes = {args.procs}, "
        f"concurrency = {args.concurrency}, "
        f"bytes = {n_bytes}, "
        f"frames = {args.frames}, "
        f"seconds = {args.seconds}, "
        f"uvloop = {args.uvloop}, "
        f"tls = {args.tls}"
    )
    asyncio.run(
        run(n_bytes, args.frames, args.seconds, args.procs, args.concurrency, args.tls)
    )


if __name__ == "__main__":
    main()
```

</details>

There are many different axis to benchmark on, instead of posting complete results I'll just summarize the current status below. If you're interested in specific numbers, I encourage you to pull this branch locally and run the benchmark yourself.

### Does uvloop help?

Kind of? It really depends on the bench flags used.

- For tornado based comms, there's a minor (negligible) performance improvement with uvloop enabled. This makes sense, since tornado is implementing its own socket handling, and only using the loop to watch sockets for events. We're not running at scale large enough in this benchmark to stress the selector code much, so the default asyncio loop is fine here. This might matter more with thousands of sockets, but for this benchmark tornado with uvloop looks about the same as without.
- For asyncio based comms, there's a noticeable performance improvement for small messages or with many connections. As message size increases / number of connections decrease the benefit of uvloop decreases. If TLS is enabled uvloop also performs measurably better (we'll get into why later). A few example benchmark runs:

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e4 -p 3 -c 4
processes = 3, concurrency = 4, bytes = 10000, frames = 1, seconds = 5, uvloop = False, tls = False
15434.4 RPS
64.79033846472814 us per request
154.344 MB/s each way
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e4 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 10000, frames = 1, seconds = 5, uvloop = True, tls = False
19144.4 RPS
52.234596017634395 us per request
191.444 MB/s each way
```

### Are the asyncio comms faster than the tornado comms (no TLS)?

Generally yes. This is especially true for small messages - as the message size increases the performance difference decreases. Here we benchmark with uvloop exclusively to show the maximum payoff. Without uvloop, asyncio yields at max 20% lower RPS (for small messages), and is slightly faster for large messages.

Some numbers are below, but in general expect the asyncio comms to be 2x faster than the tornado comms for messages <= 1e4 bytes, decreasing to equal performance for messages of 1e7 bytes.

<details>
<summary> msg size = 1e3 bytes </summary>

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e3 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 1000, frames = 1, seconds = 5, uvloop = True, tls = False
21230.2 RPS
47.10271217416699 us per request
21.2302 MB/s each way
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -b 1e3 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 1000, frames = 1, seconds = 5, uvloop = True, tls = False
9412.2 RPS
106.24508616476488 us per request
9.4122 MB/s each way
```

</details>

<details>
<summary> msg size = 1e4 bytes </summary>

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e4 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 10000, frames = 1, seconds = 5, uvloop = True, tls = False
19551.4 RPS
51.14723242325358 us per request
195.514 MB/s each way
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -b 1e4 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 10000, frames = 1, seconds = 5, uvloop = True, tls = False
8996.0 RPS
111.16051578479323 us per request
89.96 MB/s each way
```

</details>

<details>
<summary> msg size = 1e5 bytes </summary>

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e5 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 100000, frames = 1, seconds = 5, uvloop = True, tls = False
11346.8 RPS
88.13057425882188 us per request
1134.68 MB/s each way
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -b 1e5 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 100000, frames = 1, seconds = 5, uvloop = True, tls = False
6839.4 RPS
146.2116559932158 us per request
683.94 MB/s each way
```

</details>

<details>
<summary> msg size = 1e6 bytes </summary>

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e6 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 1000000, frames = 1, seconds = 5, uvloop = True, tls = False
1737.6 RPS
575.5064456721916 us per request
1737.6 MB/s each way
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -b 1e6 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 1000000, frames = 1, seconds = 5, uvloop = True, tls = False
1581.0 RPS
632.5110689437065 us per request
1581.0 MB/s each way
```

</details>

<details>
<summary> msg size = 1e7 bytes </summary>

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e7 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 10000000, frames = 1, seconds = 5, uvloop = True, tls = False
149.8 RPS
6675.567423230975 us per request
1498.0 MB/s each way
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -b 1e7 -p 3 -c 4 --uvloop
processes = 3, concurrency = 4, bytes = 10000000, frames = 1, seconds = 5, uvloop = True, tls = False
143.0 RPS
6993.006993006993 us per request
1430.0 MB/s each way
```

</details>

### Are the asyncio comms faster than the tornado comms (with TLS)?

It depends. For small messages, yes. For large messages, no.

- For messages < 1e5 bytes, asyncio (with uvloop) performs better (almost 2x better for small messages)
- 1e5 byte messages seem to be the break-even point
- For messages > 1e5 bytes, tornado performs better (1.3x higher RPS for 1e6 byte messages)

I  haven't figured out why yet, but I suspect it's unnecessary buffer copying (probably on the `write` side).

Note that if uvloop is _not_ used, then the asyncio comm performance tanks even further. This is due to a few things:

- The asyncio socket transport will copy a buffer once if the full write doesn't go through immediately in a single `socket.send` call. This is more likely to happen for large messages, which are also already more expensive to copy. In our implementation we hack around this by manually managing the socket (this is what the `_ZeroCopyWriter` class does), but this doesn't apply when TLS is enabled (for that we'd need to patch out the whole transport, rather than wrap it, something that would be doable but would take a lot more code). The uvloop transports won't copy large buffers, so if uvloop is enabled this hackery is unnecessary.
- The asyncio SSL protocol class currently also copies data more than necessary on read (it uses `data_received` rather than `get_buffer`), resulting in further slowdowns on large buffers.

With some work (mostly copying and modifying code out of asyncio in the stdlib to reduce copies) we could likely close the gap between uvloop & standard asyncio, but I don't think that work would be a good use of time.

<details>
<summary> msg size = 1e3 bytes </summary>

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e3 -p 3 -c 4 --uvloop --tls
processes = 3, concurrency = 4, bytes = 1000, frames = 1, seconds = 5, uvloop = True, tls = True
14501.4 RPS
68.95885914463432 us per request
14.5014 MB/s each way
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -b 1e3 -p 3 -c 4 --uvloop --tls
processes = 3, concurrency = 4, bytes = 1000, frames = 1, seconds = 5, uvloop = True, tls = True
8184.6 RPS
122.18068079075336 us per request
8.1846 MB/s each way
```

</details>

<details>
<summary> msg size = 1e4 bytes </summary>

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e4 -p 3 -c 4 --uvloop --tls
processes = 3, concurrency = 4, bytes = 10000, frames = 1, seconds = 5, uvloop = True, tls = True
13125.6 RPS
76.18699335649418 us per request
131.256 MB/s each way
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -b 1e4 -p 3 -c 4 --uvloop --tls
processes = 3, concurrency = 4, bytes = 10000, frames = 1, seconds = 5, uvloop = True, tls = True
7404.0 RPS
135.06212857914642 us per request
74.04 MB/s each way
```

</details>

<details>
<summary> msg size = 1e5 bytes </summary>

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e5 -p 3 -c 4 --uvloop --tls
processes = 3, concurrency = 4, bytes = 100000, frames = 1, seconds = 5, uvloop = True, tls = True
3931.2 RPS
254.37525437525434 us per request
393.12 MB/s each way
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -b 1e5 -p 3 -c 4 --uvloop --tls
processes = 3, concurrency = 4, bytes = 100000, frames = 1, seconds = 5, uvloop = True, tls = True
3996.0 RPS
250.25025025025025 us per request
399.6 MB/s each way
```

</details>
<details>
<summary> msg size = 1e6 bytes </summary>

```
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=1 python bench.py -b 1e6 -p 3 -c 4 --uvloop --tls
processes = 3, concurrency = 4, bytes = 1000000, frames = 1, seconds = 5, uvloop = True, tls = True
575.4 RPS
1737.9214459506431 us per request
575.4 MB/s each way
$ DISTRIBUTED_USE_ASYNCIO_FOR_TCP=0 python bench.py -b 1e6 -p 3 -c 4 --uvloop --tls
processes = 3, concurrency = 4, bytes = 1000000, frames = 1, seconds = 5, uvloop = True, tls = True
761.4 RPS
1313.370107696349 us per request
761.4 MB/s each way
```

</details>

### Assorted Thoughts

- With TLS disabled, asyncio comms definitely perform better across all message sizes. At best, we might expect a 2x decrease in comm latency for small messages (e.g. scheduler <-> worker coordination). However! We currently use a batching mechanism (`BatchedSend`) to batch up small messages into larger payloads, so we the number of small messages on the wire is currently low. Batching comes at a cost (higher latency, higher memory usage, more copying, consistency issues (#5481), ...), but in general makes throughput higher. We may not want to drop batching entirely, but if we move to using asyncio comms we may be able to reduce the batch interval and thus improve latency in scheduler <-> worker coordination.

- For large (1e5 bytes +) messages, TLS is slower on asyncio + uvloop than it is on tornado, and I'm not yet sure why. On fast networks this would be a performance decrease, which means we probably wouldn't want to enable asyncio comms by default until they're faster in most cases.

- None of these numbers might matter on common networks where the bandwidth may be the limiting factor. This may still matter for scheduler performance, where reducing comm CPU usage can make larger clusters more viable, or for clusters using fast networks, but for cloud computing even the tornado numbers are faster than typical bandwidth (see https://github.com/dask/distributed/issues/5258).

- The `_ZeroCopyWriter` hack is unfortunate, but I think maintainable. It doesn't support windows, but an equivalent patch for windows may be doable. I'm less enthused about implementing a full transport to speedup asyncio + TLS (without uvloop) performance, as that would be *lots* more code, but it is doable.